backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
groq
